+++ examples/change_popup_color.py
@@ -14,8 +14,7 @@ axes[1].set(title='Fancy white background')
 dc1 = datacursor(left_artist, xytext=(15, -15), bbox=None)
 
 # Make the box have a white background with a fancier connecting arrow
-dc2 = datacursor(right_artist, 
-             arrowprops=dict(arrowstyle='simple', fc='white', alpha=0.5),
-             bbox=dict(boxstyle='round,pad=0.3', fc='white', alpha=0.5))
+dc2 = datacursor(right_artist, bbox=dict(fc='white'),
+                 arrowprops=dict(arrowstyle='simple', fc='white', alpha=0.5))
 
 plt.show()